added new readonly data service for coming application; relocated middleware to shared; added try-catch for a failing person deletion attemp when addresses or telephone connections still exist + returning a proper 400 instead of return a 500

+++ src/DatabaseService/Services/PersonDataService.cs
@@ -23,6 +23,11 @@ public sealed class PersonDataService(DatabaseContext dbContext) : IPersonDataSe
         var entity = await this.GetBasePersonQuery()
             .FirstOrDefaultAsync(p => p.Id == id, cancellationToken);
 
+        if (entity?.Addresses.Count > 0 || entity?.TelephoneConnections.Count > 0)
+        {
+            throw new InvalidOperationException("Cannot delete a person with addresses or telephone connections.");
+        }
+
         await dbContext.DeleteEntity(entity, cancellationToken);
     }
